Removed share options
Share doesn't work due to a gradio bug, not sure why someone added that. Especially because queuing was broken with share even before we switched to Blocks interface (which is the reason for the share issue).
Some people are using other workarounds to share with friends so need queuing, iirc it doesn't cause any issues to have queuing enabled by default.