Changed relauncher.py to have easier to set arguments for users, added inbrowser as an argument choice
Changed `relauncher.py` to have easier to set argument parameters for users to change. No longer is editing the main `os.system`  string needed. Now, users can choose to set some of the commonly used arguments (totally open to more additions here) to be True or False, with an added area for users to enter their own custom arguments as desired.

This really should make it easier for users to toggle what parameters they wish to launch with, and prevents them from needing to edit the main launch string for the `webui.py` file.

Addition changes to `relauncher.py` include increasing the relaunch time from 0.5 seconds to 1 second, to allow for easier program closing from the Terminal.

Changed `webui.py` to now include an  `--inbrowser` argument to open the interface automatically in a user's default browser. The default is set to `False`. Gradio params have also now been alphabetised. A notice was given to users waiting for port 7860 to open that it may take upwards of 60 seconds to become available.