Enhanced support for variants
I played a lot with variants and wanted to keep track of them by
extending the filename with the variant amount and variant seed.

While doing that, I found that currently, if you give a variant seed and
an image seed and generate more than one image, you end up with the
same image for all runs.

I used increasing variants with the same variant seed for creating
movies that show how a variant is increasingly deviating. I think
those are fascinating and added functionality to WebGUI.

If you now set a seed for the image and the variant and generate more
than one image, it will increase the variant amount by the initial
amount every step. So you can easily make a series of growing variants
and have them fully reproducible.

I hope I made sense :)