add scoped seed generators in scn2img
seed generators can be initialized with "initial_seed" and will be used for this section and its children.
they can recursively used, i.e. children can do the same and the seed generator will only apply to them
and their children.
the first seed generator is initialized with the seed argument supplied to scn2img function.

this makes it possible to save initial seeds in scn2img-prompt and reuse prompts with (nearly) deterministic
results.