fix broken parameter copy in img2img, fixes #1299
changes the javascript function for copying to `js_copy_txt2img_output`, the one used for txt2img parameter copy.
this function correctly copies highlighted text, but throws an javascript error into the web console for both txt2img and img2img.
the error was caused by not returning an empty (update) list back to gradio.
to fix this the missing return value was added to the function.