Fixed Streamlit bug where RealESRGAN seemed broke
There was a safety check where RealESRGAN and loopback
were not allowed to be on at the same time to prevent
rapidly growing images (I think, at least) which was
causing confusion in the UI since there was no indication
this wasn't allowed. Using ESRGAN just on the final iteration
should be safe, so this commit enables that.