streamlit: fix whitespace in some files
Not trying to push a personal preference here, but I literally cannot
configure my editor to not fix the whitespace on individual files on a
normal save. So to prevent the inevitable scenario in which I
accidentally change the entire files whitespace mixed in with a real
change, I'm just letting it do its thing on files I'm touching in its
own commit.