Improve Docker image layer caching.
The idea here is that each set of model weights is downloaded as a separate build in a multistage build so that someone adding/re-arranging/updating model weights doesn't cause the user to have to re-download all of the weights, they will pull from cache for everything that isn't changed.

It also uses COPY --link to ensure that we get very aggressive caching of everything, even when other lines change.

I have also removed the other images since they are redundant.  They were being used as a means of caching, but we get this for free with docker and can leverage it even more with multistage builds like this.

Cross platform building is something I'm not particularly experienced with, so I have left out AMD builds from this and will need assistance from others getting that implemented.

I have a TODO in there for downloading the concepts library from GitHub, this is just me needing to figure out the right magic to wget | zip directly to the right place with some leading directories stripped out.

I also don't know where to get the ViT model weights, so that isn't included currently but I understand they are part of the currently published docker image.